+++ packages/f-searchbox/src/components/formElements/FormSearchGeo.vue
@@ -32,7 +32,6 @@ export default {
 
         service: {
             type: Object,
-            default: () => {},
             required: true
         }
     },