+++ packages/f-searchbox/src/components/formElements/FormSearchInnerFieldWrapper.vue
@@ -33,7 +33,6 @@ export default {
 
         service: {
             type: Object,
-            default: () => {},
             required: true
         }
     },